+++ src/classes/Application/Media/DocumentTrait.php
@@ -14,8 +14,12 @@ use Application_Request;
 use Application_User;
 use AppUtils\ConvertHelper;
 use AppUtils\FileHelper;
+use AppUtils\FileHelper\FileInfo;
 use AppUtils\FileHelper_Exception;
 use AppUtils\ImageHelper;
+use AppUtils\ImageHelper\ImageFormats\Formats\GIFImage;
+use AppUtils\ImageHelper\ImageFormats\FormatsCollection;
+use AppUtils\ImageHelper\ImageFormats\ImageFormatInterface;
 use AppUtils\ImageHelper_Exception;
 
 trait DocumentTrait
@@ -118,7 +122,7 @@ trait DocumentTrait
         $sourcePath = $this->getThumbnailSourcePath();
         $source = $this->getThumbnailSourceImage();
 
-        if($source->isVector()) {
+        if($this->supportsThumbnails()) {
             return $sourcePath;
         }
 
@@ -132,6 +136,29 @@ trait DocumentTrait
         }
 
         return $targetFile;
+    }
+
+    public function getImageFormat() : ImageFormatInterface
+    {
+        return FormatsCollection::getInstance()->getByExtension($this->getExtension());
+    }
+
+    public function supportsThumbnails() : bool
+    {
+        return self::formatSupportsThumbnails($this->getImageFormat(), $this->getPath());
+    }
+
+    public static function formatSupportsThumbnails(ImageFormatInterface $format, string $filePath) : bool
+    {
+        if($format->isVector()) {
+            return false;
+        }
+
+        if($format instanceof GIFImage && $format->fileHasAnimation(FileInfo::factory($filePath))) {
+            return false;
+        }
+
+        return true;
     }
 
     /**